+++ src/ispd/motor/workload/impl/TraceFileWorkloadGenerator.java
@@ -3,9 +3,9 @@ package ispd.motor.workload.impl;
 import ispd.escalonador.Escalonador;
 import ispd.motor.workload.WorkloadGenerator;
 import ispd.motor.workload.WorkloadGeneratorType;
-import ispd.motor.workload.task.ExternalTraceTaskBuilder;
-import ispd.motor.workload.task.TraceTaskBuilder;
-import ispd.motor.workload.task.TraceTaskInfo;
+import ispd.motor.workload.impl.task.ExternalTraceTaskBuilder;
+import ispd.motor.workload.impl.task.TraceTaskBuilder;
+import ispd.motor.workload.impl.task.TraceTaskInfo;
 import ispd.motor.filas.RedeDeFilas;
 import ispd.motor.filas.Tarefa;
 import ispd.motor.filas.servidores.implementacao.CS_Mestre;
