+++ src/ispd/motor/workload/impl/task/ExternalTraceTaskBuilder.java
@@ -1,4 +1,4 @@
-package ispd.motor.workload.task;
+package ispd.motor.workload.impl.task;
 
 import ispd.motor.filas.RedeDeFilas;
 import ispd.motor.filas.Tarefa;
